requirements.txt
python-docx
abiword
